Defer extent to expansion if not defined.
Now we have an inheritance like mechanism. If something is not defined
for a shape which is a template, then expand the template and see if the
method is defined for the expansion.

OO like, but not really OO.

This fits pretty nicely with the core principle that things should have
defined behaviour which can be modified by the embedding context.